Adding initial records clone script to prepare for a release
Allows a selection of datasets to be copied into new files based on the contents of the ADD core collection.

This adds a new convention for where source records will come from, now that OneDrive is removed as a records store. So far this only clones the records. It will likely be expanded to prepare them for editing and to produce a markdown table.

In time this use use the Catalogue itself but initially it uses a local store (add-remove-wms-pre-head) as a standin. Other scripts aren’t aware of this yet and will be updated in due course.